Replace hash input to field() with list of Pairs
This is because a hash does not guarantee a given order.  The reason the
simple.t and ascii.t tests worked previously is due to a quirk of MoarVM
where the input ordering was preserved on hashes.  This has since been
corrected, and hashes now behave as unordered collections similarly to other
languages (e.g. Perl 5).  In order to retain the functionality that the
field() method accepts a list of key-value pairs and generates a table with
a given order from this input data, it seemed that explicitly using a list
of Pairs to be a sensible option, which is what is implemented here.